125: Archetype and Sparse Set should have same ID to have similar interface regardless of type Changed EntityHandle to have EntityId instead of just a ulid. This technically has additional overhead because of storing an extra 64 bit uint per handle, however, this makes for easier use between EntityHandle and MutableEntityHandle. Some confusion could arise since archetypes don't technically have runtime ID, they are just unique ID, so their runtime ID would return the unique ID, but then having get unique ID return the same ID could be confusing. This change is apart of another one to obtain entity handles based off their unique ID.

+++ Source/ScarlEntProject/ScarlEnt/System.h
@@ -32,7 +32,7 @@ struct ISystem
      * @param entityIds The runtime entity identifiers of mutable entities with the subset of components.
      * @param componentSparseSet The \ref SparseSet to the component arrays.
      */
-    virtual void UpdateSparseSet(const std::unordered_set<uint32>& entityIds, SparseSet<ISparseComponentArray*, Registry::COMPONENTS_PAGE_SIZE>& componentSparseSet) = 0;
+    virtual void UpdateSparseSet(const std::unordered_set<uint64>& entityIds, SparseSet<ISparseComponentArray*, Registry::COMPONENTS_PAGE_SIZE>& componentSparseSet) = 0;
 
     uint64 componentBitset = 0;
 };
@@ -74,7 +74,7 @@ struct System : ISystem
     /**
      * @copydoc ISystem::UpdateSparseSet
      */
-    void UpdateSparseSet(const std::unordered_set<uint32>& entityIds, SparseSet<ISparseComponentArray*, Registry::COMPONENTS_PAGE_SIZE>& componentSparseSet) override
+    void UpdateSparseSet(const std::unordered_set<uint64>& entityIds, SparseSet<ISparseComponentArray*, Registry::COMPONENTS_PAGE_SIZE>& componentSparseSet) override
     {
         auto components = std::forward_as_tuple<SparseSet<Components, Registry::COMPONENT_BITSET_PAGE_SIZE>&...>(
             *static_cast<SparseSet<Components, Registry::COMPONENT_BITSET_PAGE_SIZE>*>(componentSparseSet[Registry::Instance().GetOrRegisterComponentId<Components>().id])...);
